Change property named 'precision' in `SECPercent` and finish the testcase for `SECPercent`

+++ SwiftyEcharts/Tools/Type/SECPercent.swift
@@ -9,7 +9,7 @@
 /// 用来标示百分比的情况， 通常可以通过用 数值 + % 形式来初始化
 /// 并且可以通过 Percent.precision 来进行全局定义
 public struct SECPercent {
-    public static var precision: UInt8 = 2
+    public static var precision: Int = 2
     private let value: Double
     
     public init(_ value: Double) {
